♻️(refactor) extract pip control bar component
centralize pip controls in reusable component
